+++ rs/embedders/src/wasm_utils.rs
@@ -43,7 +43,6 @@ pub struct Complexity(pub u64);
 #[derive(Debug, PartialEq, Eq, Default)]
 pub struct WasmValidationDetails {
     pub imports_details: WasmImportsDetails,
-    pub num_tables: usize,
     pub wasm_metadata: WasmMetadata,
     pub largest_function_instruction_count: NumInstructions,
     pub max_complexity: Complexity,
@@ -210,6 +209,8 @@ fn validate_and_instrument(
         config.metering_type,
         config.subnet_type,
         config.dirty_page_overhead,
+        config.max_wasm_memory_size,
+        config.max_stable_memory_size,
     )?;
     Ok((wasm_validation_details, instrumentation_output))
 }
@@ -235,7 +236,6 @@ fn compile_inner(
     let instance_pre = embedder.pre_instantiate(&module)?;
     let largest_function_instruction_count =
         wasm_validation_details.largest_function_instruction_count;
-    let num_tables = wasm_validation_details.num_tables;
     let max_complexity = wasm_validation_details.max_complexity.0;
     let serialized_module =
         SerializedModule::new(&module, instrumentation_output, wasm_validation_details)?;
@@ -245,7 +245,6 @@ fn compile_inner(
             largest_function_instruction_count,
             compilation_time: timer.elapsed(),
             max_complexity,
-            num_tables,
         },
         serialized_module,
     ))