+++ rs/nns/governance/src/gen/ic_nns_governance.pb.v1.rs
@@ -1630,7 +1630,6 @@ pub mod neurons_fund_snapshot {
         #[prost(message, optional, tag = "6")]
         pub controller: ::core::option::Option<::ic_base_types::PrincipalId>,
         /// The principals that can vote, propose, and follow on behalf of this neuron.
-        /// TODO(NNS1-3199): Populate this field with the neuron's hotkeys.
         #[prost(message, repeated, tag = "7")]
         pub hotkeys: ::prost::alloc::vec::Vec<::ic_base_types::PrincipalId>,
         /// Deprecated. Please use `controller` instead (not `hotkeys`!)
@@ -2805,6 +2804,9 @@ pub mod governance {
         #[prost(message, optional, tag = "38")]
         pub non_self_authenticating_controller_neuron_subset_metrics:
             ::core::option::Option<governance_cached_metrics::NeuronSubsetMetrics>,
+        #[prost(message, optional, tag = "39")]
+        pub public_neuron_subset_metrics:
+            ::core::option::Option<governance_cached_metrics::NeuronSubsetMetrics>,
     }
     /// Nested message and enum types in `GovernanceCachedMetrics`.
     pub mod governance_cached_metrics {
@@ -3362,8 +3364,8 @@ pub mod settle_neurons_fund_participation_request {
 /// the Neurons' Fund. However, this distribution also needs to be made available to the SNS Swap
 /// that will use this information to create SNS neurons of an appropriate size for each
 /// Neurons' Fund (as well as direct) participant. That is why in the `committed` case,
-/// the NNS Governance should populate the `neurons_fund_participants` field, while in the `aborted`
-/// case it should be empty.
+/// the NNS Governance provides `neurons_fund_neuron_portions`, while in the `aborted`
+/// case it does not.
 ///
 /// TODO(NNS1-1589): Until the Jira ticket gets solved, changes here need to be
 /// manually propagated to (sns) swap.proto.
@@ -3682,6 +3684,37 @@ pub mod restore_aging_summary {
                 _ => None,
             }
         }
+    }
+}
+/// The historical rewards that were provided to node providers, along with the contextual data
+/// needed to calculate it.
+#[derive(candid::CandidType, candid::Deserialize, serde::Serialize, comparable::Comparable)]
+#[allow(clippy::derive_partial_eq_without_eq)]
+#[derive(Clone, PartialEq, ::prost::Message)]
+pub struct ArchivedMonthlyNodeProviderRewards {
+    /// The version of the rewards data.  These versions are added to accommodate changes to the
+    /// rewards data structure over time.
+    #[prost(oneof = "archived_monthly_node_provider_rewards::Version", tags = "1")]
+    pub version: ::core::option::Option<archived_monthly_node_provider_rewards::Version>,
+}
+/// Nested message and enum types in `ArchivedMonthlyNodeProviderRewards`.
+pub mod archived_monthly_node_provider_rewards {
+    /// The first version of the stored rewards.
+    #[derive(candid::CandidType, candid::Deserialize, serde::Serialize, comparable::Comparable)]
+    #[allow(clippy::derive_partial_eq_without_eq)]
+    #[derive(Clone, PartialEq, ::prost::Message)]
+    pub struct V1 {
+        #[prost(message, optional, tag = "1")]
+        pub rewards: ::core::option::Option<super::MonthlyNodeProviderRewards>,
+    }
+    /// The version of the rewards data.  These versions are added to accommodate changes to the
+    /// rewards data structure over time.
+    #[derive(candid::CandidType, candid::Deserialize, serde::Serialize, comparable::Comparable)]
+    #[allow(clippy::derive_partial_eq_without_eq)]
+    #[derive(Clone, PartialEq, ::prost::Oneof)]
+    pub enum Version {
+        #[prost(message, tag = "1")]
+        Version1(V1),
     }
 }
 /// Proposal types are organized into topics. Neurons can automatically